Don't trim trailing whitespaces but remove a single CR at the EOL
If you `.trim()` the whole line, it will result in removal of all
trailing whitespaces which is not desirable.

Instead, we can remove only a single CR at the EOL using `.replace(/\r$/, '')`

Current IRC spec documents specifies messages to be separated with CRLF
not not LF, but most IRC servers accepts LF-separated messages so we
need to support it too.

Since `linewise` package separates a stream into LF-separated lines,
slate-irc-parser need to remove a CR at the end of the line.

Reference:
  https://tools.ietf.org/html/rfc1459#section-2.3.1
  https://tools.ietf.org/html/rfc2812#section-2.3.1